Readme: Await connection end in main
I think it is helpful to clearly communicate how to wait until the connection is closed before allowing the program to exit. (It took me some time to figure out why my application is exiting at once and how to prevent it.)

Feel free to propose a better way of doing this :-)

Thank you!